requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-genai>=1.0.0
pydantic>=2.7.0
python-dotenv>=1.0.1
langgraph>=0.0.30
langchain-core>=0.1.45
langchain-google-genai>=1.0.1
